refactor: move record batches to logstream
