+++ engine/src/main/java/io/camunda/zeebe/engine/api/EmptyProcessingResult.java
@@ -7,7 +7,7 @@
  */
 package io.camunda.zeebe.engine.api;
 
-import io.camunda.zeebe.engine.api.records.ImmutableRecordBatch;
+import io.camunda.zeebe.logstreams.ImmutableRecordBatch;
 import io.camunda.zeebe.engine.api.records.RecordBatch;
 import java.util.Optional;
