+++ engine/src/main/java/io/camunda/zeebe/engine/api/records/RecordBatch.java
@@ -7,6 +7,8 @@
  */
 package io.camunda.zeebe.engine.api.records;
 
+import io.camunda.zeebe.logstreams.ImmutableRecordBatch;
+import io.camunda.zeebe.logstreams.ImmutableRecordBatchEntry;
 import io.camunda.zeebe.protocol.record.RecordType;
 import io.camunda.zeebe.protocol.record.RejectionType;
 import io.camunda.zeebe.protocol.record.ValueType;
@@ -70,6 +72,11 @@ public final class RecordBatch implements MutableRecordBatch {
   @Override
   public boolean canAppendRecordOfLength(final int recordLength) {
     return recordBatchSizePredicate.test(recordBatchEntries.size() + 1, batchSize + recordLength);
+  }
+
+  @Override
+  public int size() {
+    return batchSize;
   }
 
   public int getBatchSize() {
