+++ logstreams/src/main/java/io/camunda/zeebe/logstreams/ImmutableRecordBatchEntry.java
@@ -5,14 +5,14 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.engine.api.records;
+package io.camunda.zeebe.logstreams;
 
 import io.camunda.zeebe.protocol.impl.record.RecordMetadata;
 import io.camunda.zeebe.protocol.impl.record.UnifiedRecordValue;
 
 /**
- * Represents an unmodifiable entry of an {@link ImmutableRecordBatch}. Contains data about a record
- * which has been created by an RecordProcessor.
+ * Represents an unmodifiable entry of a batch. Contains data about a record which has been created
+ * by an RecordProcessor.
  */
 public interface ImmutableRecordBatchEntry {
